Removed unused test controllers and fixed sonar issues
