+++ src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting.ApiTests/Controllers/RootOptionsController.cs
@@ -1,0 +1,16 @@
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
+using Spydersoft.Platform.Hosting.ApiTests.OptionsTests;
+
+namespace Spydersoft.Platform.Hosting.ApiTests.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class RootOptionsController(IOptions<RootOptionSection> rootOptions) : ControllerBase
+{
+    [HttpGet()]
+    public RootOptionSection GetRoot()
+    {
+        return rootOptions.Value;
+    }
+}
