+++ src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting.ApiTests/Controllers/UntaggedOptionsController.cs
@@ -1,0 +1,16 @@
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
+using Spydersoft.Platform.Hosting.ApiTests.OptionsTests;
+
+namespace Spydersoft.Platform.Hosting.ApiTests.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class UntaggedOptionsController(IOptions<UntaggedOptions> untaggedOptions) : ControllerBase
+{
+    [HttpGet()]
+    public UntaggedOptions GetUntagged()
+    {
+        return untaggedOptions.Value;
+    }
+}
