--- src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting.ApiTests/Controllers/WeatherForecastController.cs
@@ -1,38 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using Spydersoft.Platform.Hosting.ApiTests.Models;
-
-namespace Spydersoft.Platform.Hosting.ApiTests.Controllers;
-
-[ApiController]
-[Route("[controller]")]
-public class WeatherForecastController(ILogger<WeatherForecastController> logger) : ControllerBase
-{
-    private static readonly string[] Summaries =
-    [
-        "Freezing",
-        "Bracing",
-        "Chilly",
-        "Cool",
-        "Mild",
-        "Warm",
-        "Balmy",
-        "Hot",
-        "Sweltering",
-        "Scorching"
-    ];
-
-    private readonly ILogger<WeatherForecastController> _logger = logger;
-
-    [HttpGet]
-    public IEnumerable<WeatherForecast> Get()
-    {
-        _logger.LogInformation("Getting weather forecast");
-        return Enumerable.Range(1, 5).Select(index => new WeatherForecast
-        {
-            Date = DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
-            TemperatureC = Random.Shared.Next(-20, 55),
-            Summary = Summaries[Random.Shared.Next(Summaries.Length)]
-        })
-        .ToArray();
-    }
-}
