--- src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting.ApiTests/Models/WeatherForecast.cs
@@ -1,12 +1,0 @@
-namespace Spydersoft.Platform.Hosting.ApiTests.Models;
-
-public class WeatherForecast
-{
-    public DateOnly Date { get; set; }
-
-    public int TemperatureC { get; set; }
-
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-
-    public string? Summary { get; set; }
-}
